Move get methods in a BaseGetter class
